web/requirements.txt
Flask-Cors==2.1.0
Flask-SQLAlchemy==2.0
MySQL-python==1.2.3